pyproject.toml
[project]
name = "agentic-recruitment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "coverage>=7.13.2",
    "docling>=2.70.0",
    "google-genai>=1.60.0",
    "jinja2>=3.1.6",
    "openai>=1.82.0",
    "pydantic>=2.12.5",
]
